Remediate CodeQL/Mull findings and restructure analysis tooling
Structural fixes (Steps 1-2):
- Widen loop variables from uint8_t to size_t in SICM001A/SIVW001A (C1/C2)
- Eliminate variable shadowing in SITM001A (C4)
- Remove unused variable and lambda capture in CIAN001A/SIMT001A (Q7/Q8)
- Delete empty if-block in FMMX002A (Q1)
- Suppress unused structured binding warnings in HRRN001A/SIVD001A/TIWF001A/FMSL001A (Q2-Q4)
- Propagate transport failures in FMAL001A/002A/003A compilers (Q9)
- Handle rollback failure in mutation engine SIMT001A (C3)

Discriminating tests (Steps 3-6, 1537 -> 1545 tests):
- Beat operator>= cross-multiplication (M1)
- ScoreTime exact-end boundary (M2)
- Velocity articulation offsets (M4)
- Non-C key cadence modular arithmetic (M5)
- Empty/single-pitch analysis boundary (M6)
- Stationary voice parallel motion (M7)
- Double-digit octave parsing (M8)
- FailingTransport propagation (Q9 validation)

Analysis restructure:
- Move CodeQL queries and Mull config into analysis/ directory
- Add codeql-db/, codeql-results/, mull-build/ to .gitignore
- Remove stale build artifact directories from root

+++ src/Sunny.Core/Score/SICM001A.cpp
@@ -246,14 +246,14 @@ Result<CompiledMidiResult> compile_to_midi(const Score& score, int ppq) {
             }
 
             // Track consumed notes for tie handling
-            std::set<std::pair<std::uint64_t, std::uint8_t>> consumed;
+            std::set<std::pair<std::uint64_t, std::size_t>> consumed;
             DynamicLevel current_dynamic = DynamicLevel::mf;
 
             for (std::size_t fi = 0; fi < flat.size(); ++fi) {
                 const auto& fn = flat[fi];
                 const NoteGroup& ng = *fn.ng;
 
-                for (std::uint8_t ni = 0; ni < ng.notes.size(); ++ni) {
+                for (std::size_t ni = 0; ni < ng.notes.size(); ++ni) {
                     if (consumed.contains({fn.event_id.value, ni})) continue;
 
                     const Note& note = ng.notes[ni];
@@ -290,7 +290,7 @@ Result<CompiledMidiResult> compile_to_midi(const Score& score, int ppq) {
                             const auto& next_fn = flat[search];
                             const NoteGroup& next_ng = *next_fn.ng;
 
-                            for (std::uint8_t nni = 0; nni < next_ng.notes.size(); ++nni) {
+                            for (std::size_t nni = 0; nni < next_ng.notes.size(); ++nni) {
                                 if (consumed.contains({next_fn.event_id.value, nni})) continue;
                                 if (next_ng.notes[nni].pitch == tied_pitch) {
                                     consumed.insert({next_fn.event_id.value, nni});
